Small refactoring for Subreddit models.
